backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlmodel
aiosqlite
pydantic
pydantic-settings
pyyaml
httpx
websockets
litellm
langchain
langchain-openai
langchain-community
langgraph
langgraph-checkpoint-sqlite
bcrypt
PyJWT
pypdf
sentence-transformers
faiss-cpu
langchain-mcp-adapters
